+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - CASMCMS-8666: Remove `name` field from `/sessiontemplatetemplate` response.
+- CASMCMS-7902: Check size before loading image manifest to avoid OOM issues.
 
 ## [2.37.1] - 2025-04-09